tests/main/user-session-env: add test verifying environment variables inside the user session
There were numerous bug reports and forum posts about either $SNAP_MOUNT_DIR/bin
not being added to the $PATH, or application icons not appearing inside the
desktop environment (i.e. /var/lib/snapd/desktop missing from XDG_DATA_DIRS).
Add a test that verifies both PATH and XDG_DATA_DIRS are augmented with proper
snap related entries for users of bash and zsh.

The test assumes that the environment is set up by the user session (as in most
modern desktops). However, it does not fully cover the usage scenarios. For
instance, spawning a simple ZSH shell as a different user via `su - ..` is not
the same as starting a user session, and the only environment setup performed is
that of the shell. Thus on specific systems, eg. Ubuntu, zsh is not set up to
load /etc/profile in sh compatibility mode, and environment variables set up via
/etc/profile/snapd.sh may not be added.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>