cmd/snap-confine: detect base transitions on core16
The base staleness detector is not supported on core16 and while this
patch does not change that we now have another tool in our arsenal: base
transition detector. That detector is much simpler and can only detect
"big" changes, such as going from one base to another (but not between
revisions of the same base). Fortunately it is sufficient to address
this problem.

The same patch generalizes an existing test to verify base snap
migration in both directions, forward as well as back (back correctly
reproduces the issue reported).

Fixes: https://bugs.launchpad.net/snapd/+bug/1861901
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>