cmd/libsnap: add flags expressing UID/GID identity change
Jamie's review highlighted that seteuid and setegid differ from
setresuid in that the value -1 is not special and does not indicate
"please don't change me". What's worse, the -1 value is is actually used
for nobody and nogroup out in the wild.

To support that, use explicit flags for "change uid" and "change gid"
and adjust all call sites to handle that.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>