tests/main/interfaces-pulseaudio: use custom pulseaudio script, set kill-timeout
Use a custom pulseaudio server script to ensure that we are not unnecessarily
loading any other modules that may break the daemon startup process. Make sure
to include module-snap-policy as it's used in Ubuntu by default.

Increase exit timeout time of the daemon to 5m to make sure it does not exit too
early. However have the test execution time be limited to 3m.

Use verbose logging to help debug the test failures.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>