bootstate20: add next kernel to modeenv on setNext(), rm old on markSuccessful
This change implements adding the next kernel snap to the list in the modeenv,
so that the initramfs trusts the new kernel snap and loads the kernel modules
from the new kernel snap when we reboot.

It also implements removing the old kernel snap from the modeenv when we
successfully boot a kernel snap, so that there is only 2 kernel snaps during the
specific reboot cycle between trying a new kernel snap and switching state to
the new one after we reboot. All other times it should only contain the current,
active kernel snap specified by the kernel.efi symlink on ubuntu-boot.

Also update existing kernel snap unit tests for this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>