systemd: improve is-active check for 'failed' services
The following failure was seen in the wild:

```
+ echo 'Installing a new snapd snap'
Installing a new snapd snap
+ snap install --dangerous /var/lib/snapd/snaps/snapd_x1.snap
error: cannot perform the following tasks:
- Make current revision for snap "snapd" unavailable ([--root / is-active snapd.core-fixup.service] failed with exit status 3: failed
)
- Make snap "snapd" (unset) available to the system ([--root / is-active snapd.core-fixup.service] failed with exit status 3: failed
)
```

When a service is in failed state, `systemctl is-active`, returns 3 and prints
out `failed\n` to stderr. Extend the current checks to cover this case.

Document that `systemctl is-active` check for non-active service exits with
error code 3, but we check any non-0 exit status.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>