cmd/snap-confine: fix dead code
Coverity has flagged that the condition to do-while, namely num_read > 0
is effectively dead code becaus the loop never reaches that stage,
returning on any successful iteration of the loop.

The report is real but harmless. Switching to an infinite loop doesn't
change the behavior of this code but helps with static analysis.

https://scan7.coverity.com/doc/en/cov_checker_ref.html#static_checker_UNREACHABLE
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>