snapcraft.yaml: add comments, rename snapd part to snapd-deb
This makes it more clear where the info is adopted from, as there's also the
special magic value of `type: snapd`, so it's not clear looking at this if
`adopt-info: snapd` is also a magic value. Using snapd-deb makes this more
clear.

Also fix a shellcheck warning on the override-pull snippet, and elaborate on the
comment about not running tests as root (because they fail).

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>