tests/lib/prepare-restore.sh: reset failed state for snapd* in restore-each
Ideally we don't want to tinker with StartLimitInterval{,Sec} and
StartLimitBurst at all for the snapd services, as this is unlikely to be the
case in the wild. To make that easier, since we still restart snapd many times
for tests, at the end of each test reset the failed state so that they can be
we don't hit the restart rate limit.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>