tests/snapd-failover: install snapd on uc16 in prepare section
This simplifies the rest of the test significantly, and also lets us isolate the
case of going from core -> snapd to a different test, which will be less
complicated.

This does introduce a bit of tricky setup code, where we repack the currently
installed core snap into the snapd snap, using the current set of files from the
snapd snap on the edge channel, but there isn't currently a better way to do
this other than perhaps hard-coding the full list of snapd snap files somewhere.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>