tests/lib/snaps.sh: add helper to build snapd snap from core snap
This helper is needed for UC16 tests which use the snapd snap, as we
really want to test UC16 with the snapd snap from the spread run, not
necessarily the one on edge.

Ideally, we would not need to do this hack at image runtime, and instead
would be able to hack something up that works at image build-time, where
we already have most of this data/files available, but this is fine for
now.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>