tests/core: add back core to snapd transition failover test
This test was deleted and merged with the snapd-failover test, but
really should be it's own test, as it's got a different set of
pre-conditions and restore logic, etc.

Also this test now uses the snapd from the spread run, rather than the
one from edge/stable in the store.

Note: this test was originally tests/main/ubuntu-core-snapd-failover.
Signed-off-by: Ian Johnson <person.uwsome@gmail.com>