boot_test: add many boot robustness tests for UC20 MarkSuccessul and SetNextBoot
These tests emulate reboots during various critical points in the various
commit() functions using panic's, and ensures that we have stable boots and even
when we are rebooted during the middle of doing things.

There is a helper here which emulates the bootloader failover logic which
perhaps should move somewhere else.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>