tests: run ipv6 network-retry test too
This commit changes the retry-network test so that ipv6 is also
tested. Note that this needs an enabled "lo" interface inside
the network namespace so that ipv6 can setup a local address.
Without this we get EADDRNOTAVAIL which indicates there is no
ipv6 capable interface. It's a bit of an open question if the
EADDRNOTAVAIL should also return "NoNetwork: true" as it's
pretty specific to ipv6 it seems.