interfaces: make gpio robust against not-existing gpios in /sys
Right now the gpio interface fails with an error if a gpio device
node is missing in /sys. This leads to a not upgradable snapd.

This commit changes the code to warn only instead of failing. In
the longer run we probably want to do something smarter here but
this will unblock LP: 1866424