snap: do not hardlink on overlayfs
We currently try to hardlink a snap on install first. However this
will cause a lot of memory usage on an overlayfs based system like
the ubuntu live-cd.

To avoid this extra memory, this commit detects an overlayfs and
uses a symlink in this case instead of a hardlink.  This saves
~400Mb on the 20.04 livecd.

We could potentially just reverse the order of the hardlink/symlink
but that is a more risky change.

This should fix LP:1867415