snap-bootstrap: don't exclude partitions from sanity check
In the previous approach where we deleted all partitions with the attribute
set, regardless of being known or not (thus the sanity check ignored them
all because they would be deleted anyway). In the new approach we don't
need to exclude partitions with the attribute so this needs to be fixed.
We'll also need an exception for the growable data partition size later
in PR #8253.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>