remove main/interfaces-openvswitch-support and the supporting snap
The only AppArmor rule for openvswitch-support was moved into the
default template and that was the access being used in the test. While
openvswitch-support does use the kmod backend, we don't test this
backend in spread since we rely on the system to DTRT when we write out
/etc/modules-load.d/snap.<name>.conf (and the unit tests test that this
is written out correctly). As such, remove the spread test and its
supporting snap (we can bring this back if an AppArmor rule is added
back to the interface).