packaging,tests: ensure debian-sid builds without vendor/
The debian-sid packaging does not currently build in a pristine
sbuild environment. The nightly suite did catch this but we did
not pay attention.

This commit fixes the first issue and removes the vendor/ content
when building the debian package in debian-sid. It also updates
the packaging to exclude the "snap-bootstrap" binary which contains
code that needs a vendored package. But snap-bootstrap is not used
in debian (and never will be) so that's fine.