interfaces/desktop: limit access to /usr/share/applications
Currently, the desktop interface allows access to all of
/usr/share/applications. The files in /usr/share/applications come from
the core snaps, and with the exception of files related to xdg-open,
snaps are unable to do anything useful with these files (since they are
unable to execute the application). As such, limit the access to:

  /usr/share/applications/ r,
  /usr/share/applications/mimeapps.list r,
  /usr/share/applications/xdg-open.desktop r,

and suppress noisy denials by adding explicit deny rules for the desktop
files known to be shipped in the core, core18 and core20 snaps.

References:
https://bugs.launchpad.net/snapd/+bug/1868051 (related)