many: improve environment handling, fixing duplicate entries (#8242)
* strutil: add Environment, EnvironmentDelta and RawEnvironment

This patch adds three new types and supporting methods.

Environment is an unordered map representing environment entries that is
convenient to modify.  EnvironmentDelta is an ordered map representing
changes to an environment that also supports variable substitution and
expansion. Finally RawEnvironment is a slice representing key=value
pairs that is useful for low-level system interfaces.

Environment and EnvironmentDelta can be modified with simple map-like
methods Get, Set, Del. RawEnvironment can only be parsed to an
Environment.

The patch also contains Environment.ApplyDelta function, which correctly
applies a single delta to a given environment.

The types are designed to avoid the mistake of appending two slices
representing environment strings together. This was the original cause
of bug https://bugs.launchpad.net/snapd/+bug/1860369

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: use Environment, EnvironmentDelta and RawEnvironment

This patch replaces all ad-hoc environment handling with the three
aforementioned types. Environment is used whenever we are working
with the real complete environment (starting from OSEnvironment).
EnvironmentDelta is used to modify it according to snap-level and
application-level or hook-level overrides. Finally RawEnvironment
is only used by snap run and snap-exec, immediately before calling
execve.

This fixes incorrect handling of environment in snap-exec and also
avoids this class of bugs by making it hard, due to type system,
to express mistakes again.

Fixes: https://bugs.launchpad.net/snapd/+bug/1860369
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* osutil/env: import all symbols from check.v1

This style is shorter and we use it nearly everywhere else.
I'm about to combine more tests into this file and I don't want
to invert their style over to the verbose one.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: move new environment code to osutil

There's a lot of churn but that's unavoidable in such a move.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: remove RawEnvironment type

The raw environment is necessary only for executing programs. We can
remove the type and replace it with []string in tests (where it has most
of the usage) and a few call sites.

The corresponding method from Environment was renamed to ForExec().

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: make Environment a map[string]string

Some extra methods such as Get, Set and Delete are retained but they
are all arguably unnecessarily and will be removed in another pass.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: remove most map-like methods from Environment

This includes Get, Delete and Contains but not Set. Set has extra
semantics that will be removed in another pass.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: remove Environment.Set

This patch removes the leftover map-like method from the Environment
type. Some extra care was applied to ensure we don't try to write
through a nil map, something that Set protected against.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-exec: use NewEnvironment in test code

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: rename EnvironmentDelta to ExpandableEnv

The main feature of what was called EnvironmentDelta is the fact it
could expand expressions in variable definitions. This makes it more
obvious.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: make ExpandableEnv an immutable *strutil.OrderedMap

There's a lot of extra changes to make the journey more complete.

First of all both snap.AppInfo and snap.HookInfo provide a method
EnvStack() []osutil.ExpandableEnv which replaces the earlier
EnvironmentOverrides. Instead of merging environment entries the
entire chain is modeled. This will be relevant when we want to expand
values.

This also means that support code to make EnvStack mutable is gone.

Second of all, snap/snapenv.ExecEnv has reverted to using plain
Environment (aka map[string]string) for basicEnv and userEnv. The
values never had expressions so using the more sophisticated code
there is not really necessary.

Delta from master could be reduced further but I went for clarity
of the progression over the clarity of the end-to-end diff.

Lastly, a small but important change, snap-exec now expands each
ExpandableEnv in sequence. First for the snap-level and then for the
app or hook level. This should fix the issue that each level wanted
to expand but not completely erase the value set previously.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* osutil: remove NewEnvironment

With the current type it's not any better than just defining a map
locally.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* osutil: make ParseRawEnvironment private

It is only required for OSEnvironment and is not used anywhere outside
of the package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: rename ApplyDelta to SetExpandableEnv

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* osutil: update stale documentation referring to delta

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* osutil: weaken SetExpandableEnv

When initially implemented as a part of this branch, SetExpandableEnv
(nee ApplyDelta) behaved in somewhat unusual and "strong" way, where
forward references _were_ expanded. This was not previously so and
arguably is not needed.

As the code stands now, we have proper chaining starting from system
environment, to snap-level environment declarations (that can expand
variables) and all the way down to application or hook-level environment
declarations (also with variable expansions). Since such declarations
are ordered this provides the equivalence of shell execution performing
the same expansions.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: special-case Environment.Transform

Instead of offering specialized environment transformation functions
offer just the things we need: escaping and un-escaping of unsafe
variables removed by the dynamic linker when loading binaries with
AT_SECURE flag.

Various bits move from snap/snapenv to osutil.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: EnvStack to EnvChain, SetExpandableEnv to ExtendWithExpanded

also adjust doc comments and cover ExtendWithExpanded for the nil
initial env case

* cmd/snap,snap/snapenv: change snapenv.ExecEnv into ExtendEnvForRun

* many: make preserving ld.so setuid unsafe vars a boundary concern

support the escaping/unescaping that we use for preserving vars
stripped out by ld.so for snap-confine, via variants of
OSEnvironment/ForExec so that is happens at the process boundaries,
loading os env and setting env for exec

add/change some tests related to this

adjust comments/doc comments or add

Co-authored-by: Samuele Pedroni <pedronis@lucediurna.net>
Co-authored-by: Michael Vogt <mvo@ubuntu.com>