interfaces/network-manager: allow full access to /run/NetworkManager
There were some limitations previously in the access to
/run/NetworkManager/, to "protect" configuration of connections that
is written by netplan. But, that had a couple of issues:

1. It is perfectly legit for NM to change that connections
configuration. There can be reasons for that, and in any case NM can
disable those connections if desired. Also, NM has access to
/etc/netplan/, so the snap could do the change indirectly: apparmor
here was limiting how the interaction between applications happens,
not protecting anything really.

2. NM writes more files in /run/NetworkManager/ than just dhcp leases,
for instance /run/NetworkManager/devices/, and newer versions fail if
they cannot write them. To create the folder we need write permissions
in /run/NetworkManager/.