tests/cmd/snap-bootstrap: use dirs.RunMnt everywhere, rm MockRunMnt
This isn't really necessary if we just mock the root dir as c.MkDir(), and makes
the code somewhat simpler. It also simplifies some followup code which will use
the root dir for other things in /run/mnt/ubuntu-data/... during the tests so
having the root dir agree between the actual code and the tests is a useful
thing.

Note that we do have to change real code to not use dirs.SnapBlobDir here 
though and instead use the dirs.SnapBlobDirUnder helper, which was made to help
with this exact kind of situation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>