boot: add IsTrySnapError for distinguishing between current and try snap errors
Specifically in revisions(). We want to be able to use revisions() from a
function called from snap-bootstrap where we want to be maximally robust, and
this necessitates being able to tell if we failed to get the try snap revision
or the current snap revision. This also means that in the error case for the try
snap revision we should return the current snap regardless.

Also update the doc comment on revisions() to make this slightly non-standard
behavior clear.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>