boot: add InitramfsRunModeChooseSnapsToMount + exhaustive unit tests
InitramfsRunModeChooseSnapsToMount will be used from snap-bootstrap to replace
the logic there with a simpler call to this.

The unit tests should be exhaustive, with the exception that we don't test all
of the fallback cases when both kernel and base are requested, but we test at
least one case of both requested and one falls back each.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>