cmd/snap: Implement a "snap routine file-access" command.
This is intended for use by xdg-document-portal to let it make decisions
about whether certain paths are safe to pass to a confined process
without being proxied.

As such, false negatives are not a big problem: it just means a file may
end up being proxied when it doesn't need to be.

The command also only cares about sandbox imposed restrictions.  So it
might report that a regular user has "read-write" access to $SNAP_DATA
when file permissions mean otherwise.  xdg-document-portal can not do
anything about this, so it is best to treat the path as if it is
read/write.