github: move static checks and spread over
The chain of actions is now:

 - Static checks and unit tests
 - Spread canary (ubuntu 16.04 and core 16)
 - Spread stable (all but canary and unstable)
 - Spread unstable (unstable systems only)

I also switched to ./run-checks --static and --unit to avoid
any drift. I will explore making that faster later on.

I dropped the github go as it was conflicting with the one in the
archive. I also switched to Xenial image so that we are more like
the real thing when we are building and testing.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>