spread.yaml: use non-uefi enabled image for uc20
The current image has out-of-date packages, namely the cpio package, which are
preventing us from successfully repacking the kernel snap when we unpack it with
unmkinitramfs, which calls cpio directly.

Additionally, all of the default ubuntu images are moving to use uefi by
default, so we don't need to specify this anymore.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>