snap: improve TestWaitRecovers test
The SnapOpSuite.TestWaitRecovers test recently failed during the
package build. It seems to be some sort of race, the failure is:
```
----------------------------------------------------------------------
FAIL: cmd_snap_op_test.go:140: SnapOpSuite.TestWaitRecovers

cmd_snap_op_test.go:158:
    // we got the change
    c.Assert(chg, check.NotNil)
... value *client.Change = (*client.Change)(nil)
```

This commit improves the test slightly by making the error visible
instead of hidding it. It also removes snap.MockMaxGoneTime()
because that is not really useful in this test - the test is about
reconnecting after an error and setting this value looks more like
cargo-culting as it will not play a role here.