cmd/snap: don't wait for system key when stopping
When the system is shutting down waiting for system key is futile. While
we should not have allowed for the core / snapd snaps to be unmounted
before all the services have correctly stopped, fixing that is a longer
journey.

Meanwhile, we can just check if system is shutting down and if so,
ignore any system key mismatches that have occurred.

Fixes: https://bugs.launchpad.net/snapd/+bug/1871652
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>