data: snapd.apparmor.service requires mounts for
Add RequiresMountsFor= for the two directories snapd.apparmor.service
actually depends on, the location of the profiles and the location of
the cache.

This was always the case, but it slipped below our radar. This is now
relevant due to experimental ZFS support available in Ubuntu 20.04,
which puts /var/lib on a separate mount point that is not available on
early boot.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>