interfaces/apparmor: use differently templated policy for non-core bases
For snaps that specify a base for anything other than core snaps, use a
default policy that allows access to the files in the base. To achieve
this, update the template to:

- refactor common rules for all bases into a single variable
- adjust the default template to use the common rules and any additional
  rules specific to core snaps.
- create a default base template that uses the common rules as well as
  non-hostfs file rules for accessing the files in the base

Recall that as part of the mount namespace setup, some directories from
the host will be bind mounted onto the base snap (these are defined by
snap-confine). The locations of the target mounts that the snap sees at
runtime are (for clarity, not all subdirectories are listed (eg,
/var/lib/snapd/hostfs is not listed since /var/lib/snapd is)):

- /dev
- /etc
- /home
- /lib/modules and /usr/lib/modules
- /lib/firmware and /usr/lib/firmware
- /mnt, /media and /run/media
- /proc
- /root
- /run
- /snap and /var/snap
- /sys
- /usr/lib/snapd
- /usr/src
- /var/lib/dhcp
- /var/lib/extrausers
- /var/lib/jenkins
- /var/lib/snapd
- /var/log
- /var/tmp

For files coming from the host in this manner, accesses should be common to
all bases, either via the template or interface rules (eg, given the same
connected interfaces, access to devices in /dev should generally be the
same, regardless of whether the snap specifies 'base: core18' or
'base: other').

The preamble and default accesses common to all bases go in the common
rules. These rules include the aformentioned host file rules as well as
non-file rules (eg signal, dbus, unix, etc).

For clarity and ease of maintenance of base template rules, we will
whitelist top-level directories instead of using glob rules (we can add
more if specific bases dictate).

In addition to the tests added to ensure common rules are always used
with default and base default rules appropriately used, add a
TestTemplateRulesInCommon() unit test for (attempting to) detect when
common patterns of rules are added to the default template or base
default template when they should instead be in common rules.

References:
https://forum.snapcraft.io/t/base-runtime-freedesktop-sdk-runtime-19-08/11153