boot: add FindPartitionUUIDForBootedKernelDisk + tests
This will be used from snap-bootstrap initramfs-mounts when we have no
partitions mounted, as the first partition we mount should be the one that's the
same as the kernel we booted from. When we have EFI variables, then we can use
this EFI variable to figure out the UUID of the partition we booted from.

If we don't have EFI, then we should eventually do something, but for now leave
it as a TODO.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>