boot/bootstate20: re-factor kernel methods to use new interface for state
We don't want to carry around the ExtractedRunKernelImageBootloader in
bootState20Kernel because eventually we want to have alternate implementations
that don't support ExtractedRunKernelImageBootloader be used with bootstate20.

This has a slight change in behavior, in that we will call DisableTryKernel more
than we did before, but otherwise is the same in external behavior.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>