boot/bootstate20: fix bug in try-kernel cleanup
We should have always been disabling the try-kernel in MarkBootSuccessful here,
but we weren't because in this specific test scenario, trySnap was nil, because
TryKernel() didn't return a real snap, but the status was TryingStatus, so we
chose the try-kernel (which was nil) as the snap to use as the booted snap.

The code in commit() for bootState20MarkSuccessful was robust enough though that
it didn't panic, because it's written in such a way that if the booted kernel
snap was nil, it will just skip some of the cleanup there. That's so that we can
call MarkSuccessful on just a single snap type, such as base or kernel only, but
in reality we will always call it on both snaps from boot.MarkBootSuccessful.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>