snap-bootstrap: lock access to sealed keys
We should lock access to sealed keys whenever TPM is available, regardless
of whether secure boot is enabled or there is a volume to unlock. Note that
snap-bootstrap can be called several times during initialization, and if
there are multiple volumes to unlock we should lock access to the sealed
keys only after the last encrypted volume is unlocked.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>