tests: ignore user-12345 slice and service
The tests/main/snap-confine/undesired-user-mode-group test is randomly
picking up the legitimate user session cgroup left-overs that should not
be measured here.

The interesting finding of this patch is that there are two general
forms, the one we already captured, with the user@12345.service earlier
in the chain and then the one where user-12345.slice is the parent of
the user@12345.service, oh well, systemd just changing things along the
way.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>