cmd/snap-bootstrap/initramfs-mounts: support EnvRefExtractedKernelBootloader's
This supports bootloaders like u-boot that don't implement
ExtractedRunKernelImageBootloader, where we can only use the bootenv variables
directly for referencing kernels.

Also add unit tests around this kind of bootloader, but note that these tests
are very duplicative, but we can delete the duplicative tests here when we move
all of this logic around which snap to mount to the boot pkg where we will be in
a much better place to test the additional dimension we are adding here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>