boot/makebootable: write boot vars after EnableKernel to share code
Also update the doc-comments about the ordering here. The ordering of enabling
the kernel.efi symlink and writing the ubuntu-boot bootenv vars doesn't actually
matter since the bootloader won't look at ubuntu-boot until we write
snapd_recovery_mode=run on the ubuntu-seed bootloader, so as long as that
happens last we are okay.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>