cmd/snap-bootstrap/initramfs-mounts: return after mounting ubuntu-{boot,seed}
This is necessary because ubuntu-seed will contain the keyfile we need
to unlock ubuntu-data-enc in the encrypted case.

Also we re-order this to mount ubuntu-boot because we will eventually
need to be smarter about mounting ubuntu-boot to match the booted kernel
which will require mounting ubuntu-boot first before anything else.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>