overlord/snapstate: warn of refresh/postpone events
When refresh-app-awareness, aka tracking of active applications first
prevents a snap from refreshing a warning is issued to alert the user
that the snap will refresh forcibly after a week.

When the time elapses and force-refresh is issued a second warning
alerts the user of this behavior.

This is done in the hope that the desktop session will eventually
surface those messages, allowing the users to act accordingly. In
addition, the presence of the second warning may assist in debugging of
data-loss like situations during a forced refresh.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>