packaging/fedora: disable FIPS compliant crypto for static binaries
Since 1.12.1-1, the Go toolchain available in go-toolset module for RHEL8
defaults to the FIPS compliant boringcrypto implementation. This functionality
dynamically loads libcrypto.so via dlopen(). This breaks statically built
binaries, since the dlopen() call is sill included, so even though the binary is
static, it stil tries to load libcrypto.so at runtime.

Disable the boringcrypto FIPS tweaks for static binaries on RHEL 8+.

[1]. https://developers.redhat.com/blog/2019/06/24/go-and-fips-140-2-on-red-hat-enterprise-linux/

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>