tests: port selinux-clean to session-tool
When running interactively I noticed this test fails to run "cmd"
correctly, resulting in this error:

    /snap/test-snapd-desktop/x1/bin/cmd: 4: exec: sh -c 'echo hello world': not found

I rewrote that fragment to be less mysterious around quoting.

This patch also contains a pair of fixes for session-tool related to SELinux.
The first issue is incorrect context on the directory
/var/lib/systemd/linger created to work around a systemd bug.
The second issue is the incorrect service context that we get from
StartTransientScope, that is not representative of the context a user
would get in a session.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>