tests: session-tool --restore -u stops user-$UID.slice
This patch expands session-tool to perform a more extensive restore
operation for non-root users. The primary motivation is porting of
tests/main/user-mounts, with an added check that the XDG_RUNTIME_DIR
directory is no longer present after a call to session-tool --restore,
where it was observed that, at least on Ubuntu 19.10 the directory would
not be removed, (since it is a mount point, it would also remain
mounted.

For unknown reasons systemd --user remains running after a call to
"loginctl disable-linger" on that system, this does not happen on Ubuntu
20.04 anymore, though, suggesting it is just another bug that was
fixed.

With this patch, when restoring a non-root user, see the code comment
for the rationale of this exception, we stop the user-UID.slice as well.
The aforementioned slice contains user@UID.session as well as any
session-NNN.scope units belonging to the user.

In addition, as a sanity check, session-tool --restore fails immediately
if the XDG_RUNTIME_DIR directory exists after the slice is stopped.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>