tests: run smoke test with different bases
The smoke suite should also run a simple echo test on
different bases. This should catch the error we had
on armhf where with kernel 5.4 certain syscalls are
not available. See:
https://github.com/snapcore/core20/issues/48

Because the smoke suite is run with autopkgtest on
all clasic ubuntu architectures a smoke test like
this would have found the issue.