tests: fix racy check for pulseaudio activity
The test code contained "pulseaudio --check" which looks for the
pulseaudio PID in /run/user/$UID/pulse/pid and then checks if that
process is alive.

This only worked because we ran "session-tool -u test --prepare"
immediately before, and pulse has just started (asynchronously),
so it was still running while we looked.

For reasons I did not investigate, in our test session, pulse stops as
soon as it is idle, which is immediately after we start the session.
Subsequent use of paplay and other test executables cause socket
activation, so in actual testing pulse *is* working.

This fixes random failures related to those two tests.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>