many: add InstallModeHostWritableDir, use during install mode
The usage of InitramfsWritableDir from userspace snapd during install mode was
unfortunate and actually ended up shadowing the tmpfs /run/mnt/ubuntu-data that
was mounted there during the initramfs, but it just happened to work. Now that
the initramfs doesn't mount /run/mnt/ubuntu-data anymore, we need to be more
explicit about which directory layout we are using during which parts.

During the initramfs, all we care about is that the system we are booting will
have it's rootfs data at /run/mnt/data, whether that is a tmpfs or a physical
partition.

During install mode, we only want to handle the physical ubuntu-data partition,
which will be mounted at /run/mnt/ubuntu-data by snapd and will be persistent
and used during run mode as /run/mnt/data.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>