cmd/snap-bootstrap: fix typo, adjust comments
When we are writing the modeenv from the initramfs, it will always be to a
tmpfs, as the real persisted modeenv is only written to the physical ubuntu-data
partition, which in install mode initramfs doesn't exist yet. In recover mode,
it is potentially dangerous to write to the physical partition, so instead we
write the recover mode modeenv to a tmpfs.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>