tests/lib: systems can have a user instance of systemd without D-Bus
"session-tool --prepare" was only trying to start default.target if it
believed there was a systemd controlled D-Bus session bus.  This is not
the case on Ubuntu 16.04, yet it still has a working user instance of
systemd.

Instead try running "systemctl --user is-enabled default.target", which
should only succeed if we can talk to the user instance of systemd.