secboot: append uuid to ubuntu-data when decrypting
This will cause systemd-cryptsetup to mount ubuntu-data-enc at a mostly
unpredictable location in /dev/mapper.
Then requesting that we mount this node in /dev/mapper/ as the filesystem will
prevent an attacker from attaching a disk with an unencrypted partition named
ubuntu-data to the system and having snap-bootstrap mistakenly unlock the real
ubuntu-data-enc but mount the attacker's unencrypted ubuntu-data partition as
the real rootfs.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>