cmd/snap-bootstrap/initramfs-mounts, tests: re-name the steps
The steps now are logical stopping points where we return from snap-bootstrap to
"the-tool", and thus are "blocking" points where we need to wait for the mount
to complete before continuing.

Doing this now will simplify future changes and makes it easier to explain how
many times snap-bootstrap should be called.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>