cmd/snap-bootstrap/initramfs-mounts: use booted kernel partition uuid if avail
If we have information on which partition uuid the kernel was booted from, then
we should use that to mount the first partition, i.e. ubuntu-seed in the
recover/install cases, and ubuntu-boot in the run mode case.

Also add tests around this and a TODO about the next step, which is to
use this mountpoint to decide which disk/partition we should mount the next
partition from in the run mode case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>