tests/mount-ns: stop binfmt_misc mount unit
Systemd creates an automount unit for the /proc/sys/fs/binfmt_misc
filesystem.  Any non-special interaction with that directory will
trigger auto-mount behavior.  Since this is extremely easy to do, we
don't want to affect our measurement by it in any way. A simple way to
avoid that is to stop the *mount* unit, it will still be auto-mounted on
demand but will no longer be mounted while we measure.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>