osutil: add disks pkg for associating mountpoints with disks/partitions
These methods are meant to be used by snap-bootstrap, primarily to take an
already mounted partition, and use that to find the underlying disk, then to
find an unmounted partition from the same disk, with the specified label so that
we can mount that one.

Eventually DiskFromMountPoint and MountPointIsFromDisk will support "following"
decrypted mapper partitions to the encrypted physical partition with the Options
struct.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>