tests: port interfaces-contacts-service to session-tool
Testing session stuff before session-tool was really the dark ages.
Yikes! All the terrible boilerplate and hacks are gone now. I also
cleaned up the system list, so that it's easier to spot the pattern of
this interface being in real trouble of becoming useless.

As a note, dbus.sh always leaks a dbus-daemon around, so one less of
those.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>